Configure subcommand argparsers in command.py
Submodules are now lazily imported, when using the manubot CLI.
