+++ manubot/process/process_command.py
@@ -1,39 +1,4 @@
 import logging
-import pathlib
-
-
-def add_subparser_process(subparsers):
-    parser = subparsers.add_parser(
-        name='process', help='process manuscript content',
-        description='Process manuscript content to create outputs for Pandoc consumption. '
-                    'Performs bibliographic processing and templating.',
-    )
-    parser.add_argument(
-        '--content-directory', type=pathlib.Path, required=True,
-        help='Directory where manuscript content files are located.',
-    )
-    parser.add_argument(
-        '--output-directory', type=pathlib.Path, required=True,
-        help='Directory to output files generated by this script.',
-    )
-    parser.add_argument(
-        '--template-variables-path', action='append', default=[],
-        help='Path or URL of a JSON file containing template variables for jinja2. '
-             'Specify this argument multiple times to read multiple files. '
-             'Variables can be applied to a namespace (i.e. stored under a dictionary key) '
-             'like `--template-variables-path=namespace=path_or_url`. '
-             'Namespaces must match the regex `[a-zA-Z_][a-zA-Z0-9_]*`.',
-    )
-    parser.add_argument(
-        '--cache-directory', type=pathlib.Path,
-        help='Custom cache directory. '
-             'If not specified, caches to output-directory.',
-    )
-    parser.add_argument(
-        '--clear-requests-cache', action='store_true',
-    )
-    parser.set_defaults(function=cli_process)
-    return parser
 
 
 def cli_process(args):
